// src/api-chain/model/apiResponsePaginatedListChainUserCustomer.ts
export interface ApiResponsePaginatedListChainUserCustomer { 
    data?: PaginatedListChainUserCustomer;
    /**
     * Optional details for unexpected error responses.
     */
    errorDetails?: string;
    /**
     * Simple message to explain client developers the reason for error.
     */
    errorMessage?: string;
    status: ApiDefaultResponseStatusEnum;
    validationErrorDetails?: ApiValidationErrorDetails;
}

/**
 * Namespace for property- and property-value-enumerations of ApiResponsePaginatedListChainUserCustomer.
 */
export namespace ApiResponsePaginatedListChainUserCustomer {
    /**
     * All properties of ApiResponsePaginatedListChainUserCustomer.
     */
    export enum Properties {
        data = 'data',
        /**
         * Optional details for unexpected error responses.
         */
        errorDetails = 'errorDetails',
        /**
         * Simple message to explain client developers the reason for error.
         */
        errorMessage = 'errorMessage',
        status = 'status',
        validationErrorDetails = 'validationErrorDetails'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiResponsePaginatedListChainUserCustomer',
            vars: [
                {
                    metadata: PaginatedListChainUserCustomer.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'data',
                    classname: 'ApiResponsePaginatedListChainUserCustomer',
                    dataType: 'PaginatedListChainUserCustomer',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'PaginatedListChainUserCustomer'
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'errorDetails',
                    classname: 'ApiResponsePaginatedListChainUserCustomer',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'errorMessage',
                    classname: 'ApiResponsePaginatedListChainUserCustomer',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    metadata: ApiDefaultResponseStatusEnum.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'status',
                    classname: 'ApiResponsePaginatedListChainUserCustomer',
                    dataType: 'ApiDefaultResponseStatusEnum',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'ApiDefaultResponseStatusEnum'
                },
                {
                    metadata: ApiValidationErrorDetails.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'validationErrorDetails',
                    classname: 'ApiResponsePaginatedListChainUserCustomer',
                    dataType: 'ApiValidationErrorDetails',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'ApiValidationErrorDetails'
                },
            ],
            validators: {
                data: [
                ],
                errorDetails: [
                ],
                errorMessage: [
                ],
                status: [
                        ['required'],
                ],
                validationErrorDetails: [
                ],
            }
        }
    }

  // export const ApiResponsePaginatedListChainUserCustomerValidationScheme = {
  //     validators: [],
  //     fields: {
  //               data: {
  //                   validators: []
  //               },
  //               errorDetails: {
  //                   validators: []
  //               },
  //               errorMessage: {
  //                   validators: []
  //               },
  //               status: {
  //                   validators: []
  //               },
  //               validationErrorDetails: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiResponsePaginatedListChainUserCustomer>;


}